Improve label encoding structure
Previously, we had LabelEncoding structs in Config, and the rest of the program just dealt with codes. Now LabelEncoding produces a Label, which not only has the code, upon which it performs hashing and equality, but it also contains an Rc to the label text, which it automatically displays in Debug and Display.