Add optional and zero-or-more to environment
I have added the ability to make environment atoms optional, or
zero-or-more, marked with `?` and `*` respectively. It is also fully
recursive, if you need that for some reason. Furthermore, I have changed
rule application to only do one sweep, and when it applies a rule, it
goes again, starting at the next character. This is to prevent infinite
loops and stack overflows. I've also restructured the consonant
character definitions in the example file.