Implement CLI evolve subcommand
I implemented the evolve subcommand for the cli. In order to do this, I
also made the FileLexer struct more abstracted and able to lex anything
that implements BufRead. I then moved the file opening and reading code
to the parser, and then added the ability to parse words given on the
command line.