Centralize IO errors and add path to FilePosition
Now, all IO errors are handled in the same place. I also added the file
path to the FilePosition struct. By not being able to derive Copy
anymore, it forced me to rethink some memory management stuff which is
cool. By having files in the position like this, we'll be able to
implement muli-file config files in the future! It also just makes a lot
of CLI stuff cleaner. Also error reporing is much clearer and prettier.