Implement diacritic-based phoneme display
So, now rebasing a phoneme will first select a new base best base
character, like normal, but then it will take those extra attributes,
scan through the available diacritics, and apply the ones that it can
find. Unfortunately, the order in which it applies diacritics is
completely random, so I need to fix it.