Change the wording of some output flags
To make things more clear and direct, I have renamed `--show_rules` to
`--display-rules`, `--show-all-rules` to `--all-results`, and
`--show-changes` to `--changes`. I have also moved the short flag for
`--config` to `-C`, and then added `-d` for `--display-rules`, `-a` for
`--all-results`, and `-c` for `--changes`. I also changed the phrasing
of the rules for stress movement in Early PGerm so only words that need
it will have it applied.