Allow specifying type of error type in `Error`
Previously, `Error` just stored its internal type as `dyn ErrorType`,
which meant that we couldn't implement `PartialEq` for `Error`, because
`PartialEq` is not `dyn` compatible. As a result, we were unable to unit
test the data in an `Error`. I have changed it so that you can pass a
specific type that implements `ErrorType`, with `dyn ErrorType` just
being the default if none is specified. Then, if the passed `ErrorType`
implements `PartialEq`, it will automatically be implemented for the
corresponding `Error`, allowing us to unit test results if we know what
type they will return while retaining the flexibility of `dyn ErrorType`
when we don't know.