Separate `OutputAttribute` from `Attribute`
Selector codes can only be used as attributes in the output of a rule,
but the old type system let them be used anywhere. I have created a new
attribute type `OutputAttribute` which can encode selector codes, and
then removed them from the normal `Attribute`, so the compiler can be
guaranteed that it won't find selector codes outside of outputs.