Implement filter and selector parsing
This was pretty easy to do using the tools that I had already created.
Because selectors need to know their code, but the atom parsers don't
have their start tokens, so it'd need to be passed in the function
argument. I just implemented a function `try_parse` directly on
`Selector` so that it can blend in.