Implement full environment parsing
Now we can parse full environments, assuming that they exist. I had to
make the `pos` field on `Error` public so that we can find the position
to report in case of an eof before a target, but that's fine. The
coverage is starting to report more of those ghost instantiations but
I don't really know how to fix that.