Implement rule parsing
Now you can parse rules. I also changed parse_iter so that now it takes
a fallback position, and Parse also provides tracked_parse_iter which
will return the position of the last encountered token. Additionally,
the environment parser will automatically handle rules having no
environment, which means it needs to be called on the /, not on the
environment atoms.